+++ src/styles/app/14-welcome.css
@@ -1,10 +1,31 @@
 .welcomeEmptyState {
+	--welcome-page-bg: var(--bg-primary);
+	--welcome-page-muted: color-mix(in srgb, var(--text-primary) 46%, transparent);
+	--welcome-brand-bg: color-mix(in srgb, var(--text-primary) 94%, black);
+	--welcome-brand-text: var(--bg-primary);
+	--welcome-brand-muted: color-mix(in srgb, var(--bg-primary) 52%, transparent);
+	--welcome-brand-heading: color-mix(
+		in srgb,
+		var(--bg-primary) 66%,
+		var(--text-secondary)
+	);
+	--welcome-brand-grid: color-mix(in srgb, var(--bg-primary) 5%, transparent);
+	--welcome-brand-line: color-mix(in srgb, var(--bg-primary) 10%, transparent);
+	--welcome-brand-accent: color-mix(
+		in srgb,
+		var(--interactive-accent) 76%,
+		var(--status-success-fg, var(--text-primary))
+	);
+	--welcome-action-bg: var(--text-primary);
+	--welcome-action-bg-hover: color-mix(in srgb, var(--text-primary) 88%, black);
+	--welcome-action-fg: var(--bg-primary);
+
 	flex: 1;
 	width: 100%;
 	min-height: 0;
 	display: grid;
 	grid-template-columns: minmax(0, 1fr) minmax(0, 1fr);
-	background: #f3f3f1;
+	background: var(--welcome-page-bg);
 	cursor: default;
 	overflow: hidden;
 	padding: 0;
@@ -19,7 +40,7 @@
 	justify-content: center;
 	overflow: hidden;
 	padding: clamp(20px, 4vw, 72px);
-	background: #f3f3f1;
+	background: var(--welcome-page-bg);
 }
 
 .welcomeBrandContent {
@@ -30,21 +51,23 @@
 	padding: clamp(28px, 4.4vw, 78px);
 	overflow: auto;
 	border-radius: 14px;
-	background: #030303;
-	color: #f5f5f5;
-	scrollbar-width: none;
-}
-
-.welcomeBrandContent::-webkit-scrollbar {
-	display: none;
+	background: var(--welcome-brand-bg);
+	color: var(--welcome-brand-text);
+	scrollbar-width: thin;
+	scrollbar-color: color-mix(
+			in srgb,
+			var(--welcome-brand-text) 32%,
+			transparent
+		)
+		transparent;
 }
 
 .welcomeBrandContent::before {
 	content: "";
 	position: absolute;
 	inset: 0;
-	background: linear-gradient(rgba(255, 255, 255, 0.04) 1px, transparent 1px),
-		linear-gradient(90deg, rgba(255, 255, 255, 0.04) 1px, transparent 1px);
+	background: linear-gradient(var(--welcome-brand-grid) 1px, transparent 1px),
+		linear-gradient(90deg, var(--welcome-brand-grid) 1px, transparent 1px);
 	background-size: 96px 96px;
 	opacity: 0.5;
 }
@@ -56,7 +79,7 @@
 	top: 14%;
 	width: 38%;
 	height: 1px;
-	background: rgba(255, 255, 255, 0.09);
+	background: var(--welcome-brand-line);
 }
 
 .welcomeCornerGlyph,
@@ -80,7 +103,7 @@
 
 .welcomeKicker {
 	margin: 0 0 clamp(18px, 2.4vw, 32px);
-	color: #48bd67;
+	color: var(--welcome-brand-accent);
 	font-size: clamp(12px, 1.1vw, 14px);
 	font-weight: 650;
 }
@@ -88,7 +111,7 @@
 .welcomeHeading {
 	max-width: 560px;
 	margin: 0 0 clamp(26px, 3.4vw, 46px);
-	color: #a7a7a7;
+	color: var(--welcome-brand-heading);
 	font-size: clamp(34px, 4.6vw, 78px);
 	font-weight: 650;
 	letter-spacing: 0;
@@ -121,14 +144,14 @@
 	top: 3px;
 	width: 8px;
 	height: 4px;
-	border-bottom: 1.5px solid #f5f5f5;
-	border-left: 1.5px solid #f5f5f5;
+	border-bottom: 1.5px solid var(--welcome-brand-text);
+	border-left: 1.5px solid var(--welcome-brand-text);
 	transform: rotate(-45deg);
 }
 
 .welcomeFeature h2 {
 	margin: 0 0 6px;
-	color: #f5f5f5;
+	color: var(--welcome-brand-text);
 	font-size: clamp(13px, 1.2vw, 15px);
 	font-weight: 650;
 	line-height: 1.35;
@@ -136,7 +159,7 @@
 
 .welcomeFeature p {
 	margin: 0;
-	color: rgba(245, 245, 245, 0.5);
+	color: var(--welcome-brand-muted);
 	font-size: clamp(12px, 1.1vw, 14px);
 	line-height: 1.5;
 }
@@ -146,7 +169,7 @@
 	height: clamp(18px, 3vw, 40px);
 	margin: clamp(22px, 3vw, 42px) 0 0;
 	background-image: radial-gradient(
-		rgba(255, 255, 255, 0.18) 1px,
+		color-mix(in srgb, var(--welcome-brand-text) 18%, transparent) 1px,
 		transparent 1px
 	);
 	background-size: 8px 8px;
@@ -159,14 +182,14 @@
 	align-items: center;
 	justify-content: center;
 	padding: clamp(20px, 4vw, 72px);
-	background: #f3f3f1;
+	background: var(--welcome-page-bg);
 }
 
 .welcomeActionHint {
 	position: absolute;
 	left: calc(50% - clamp(18px, 2vw, 34px));
 	top: calc(50% - clamp(194px, 17vw, 262px));
-	color: rgba(36, 36, 36, 0.42);
+	color: color-mix(in srgb, var(--welcome-page-muted) 82%, transparent);
 	transform: rotate(8deg);
 	pointer-events: none;
 }
@@ -174,7 +197,7 @@
 .welcomeActionHint span {
 	display: block;
 	margin: 0 0 4px 2px;
-	color: rgba(36, 36, 36, 0.52);
+	color: var(--welcome-page-muted);
 	font-family: "Comic Sans MS", "Bradley Hand", cursive;
 	font-size: clamp(13px, 1.4vw, 18px);
 	font-weight: 600;
@@ -191,9 +214,9 @@
 .welcomeActionBtn {
 	width: min(220px, 100%);
 	min-height: 42px;
-	color: #ffffff;
-	background: #242424;
-	border: 1px solid #242424;
+	color: var(--welcome-action-fg);
+	background: var(--welcome-action-bg);
+	border: 1px solid var(--welcome-action-bg);
 	padding: 10px 24px;
 	border-radius: var(--radius-md);
 	font-size: 13px;
@@ -205,9 +228,9 @@
 }
 
 .welcomeActionBtn:hover {
-	color: #ffffff;
-	background: #1a1a1a;
-	border-color: #1a1a1a;
+	color: var(--welcome-action-fg);
+	background: var(--welcome-action-bg-hover);
+	border-color: var(--welcome-action-bg-hover);
 }
 
 .welcomeActionBtn:disabled {
